fix: temprally byte -> inte

+++ src/main/java/com/nhnacademy/bookstore/user/member/entity/Member.java
@@ -48,7 +48,7 @@ public class Member extends Customer {
 
     private LocalDate birthday;
 
-    private Byte birthdayMonth;
+    private Integer birthdayMonth;
 
     @Column(nullable = false)
     private LocalDate joinDate;
@@ -95,7 +95,7 @@ public class Member extends Customer {
         this.nickname = requestDto.nickName();
         this.gender = requestDto.gender();
         this.birthday = requestDto.birthday();
-        this.birthdayMonth = (byte) requestDto.birthday().getMonthValue();
+        this.birthdayMonth = requestDto.birthday().getMonthValue();
         this.joinDate = LocalDate.now();
         this.isPaycoLogin = false;
         this.point = 0;
